hunter integration and refactor
