+++ requirements.txt
@@ -1,14 +1,47 @@
+astroid==3.0.2
 certifi==2023.11.17
+charset-normalizer==3.3.2
 click==8.1.7
+dateparser==1.2.0
+DateTime==5.4
+decorator==5.1.1
+dill==0.3.7
 elastic-transport==8.11.0
 elasticsearch==8.11.1
 elasticsearch7==7.13.0
+expandvars==0.6.5
+gevent==23.9.1
+greenlet==3.0.3
+hunter @ git+https://github.com/datastax-labs/hunter.git@8ff166979d000780ad548e49f006ef2a15d54123
+idna==3.6
+isort==5.13.2
+mccabe==0.7.0
+more-itertools==8.14.0
+numpy==1.24.0
 fmatch==0.0.4
-numpy==1.26.3
 pandas==2.1.4
+platformdirs==4.1.0
+pylint==3.0.3
+pystache==0.6.5
 python-dateutil==2.8.2
 pytz==2023.3.post1
 PyYAML==6.0.1
+regex==2023.12.25
+requests==2.31.0
+ruamel.yaml==0.17.21
+ruamel.yaml.clib==0.2.8
+scipy==1.12.0
+signal-processing-algorithms==1.3.5
 six==1.16.0
+slack_sdk==3.26.2
+structlog==19.2.0
+tabulate==0.8.10
+tomlkit==0.12.3
+typed-ast==1.5.5
+typing-extensions==3.10.0.2
 tzdata==2023.4
+tzlocal==5.2
 urllib3==1.26.18
+validators==0.18.2
+zope.event==5.0
+zope.interface==6.1
